+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 2025-10-25
 
+- Add animation system.
+
+## 2025-10-25
+
 - Separate the preview and editor to different pages.
 
 ## 2025-10-24